Remove CommonColors and move settings into HighwayPreset class.
