+++ YARG.Core/Game/Presets/ColorProfile.cs
@@ -32,8 +32,7 @@ namespace YARG.Core.Game
         public FiveLaneDrumsColors FiveLaneDrums;
         [SettingSubSection]
         public ProKeysColors ProKeys;
-        [SettingSubSection]
-        public CommonColors Common;
+
 
         public ColorProfile(string name, bool defaultPreset = false) : base(name, defaultPreset)
         {
@@ -41,7 +40,6 @@ namespace YARG.Core.Game
             FourLaneDrums = new FourLaneDrumsColors();
             FiveLaneDrums = new FiveLaneDrumsColors();
             ProKeys = new ProKeysColors();
-            Common = new CommonColors();
         }
 
         public override BasePreset CopyWithNewName(string name)
@@ -52,7 +50,6 @@ namespace YARG.Core.Game
                 FourLaneDrums = FourLaneDrums.Copy(),
                 FiveLaneDrums = FiveLaneDrums.Copy(),
                 ProKeys = ProKeys.Copy(),
-                Common = Common.Copy()
             };
         }
 
@@ -65,7 +62,6 @@ namespace YARG.Core.Game
             FourLaneDrums.Serialize(writer);
             FiveLaneDrums.Serialize(writer);
             ProKeys.Serialize(writer);
-            Common.Serialize(writer);
         }
 
         public void Deserialize(BinaryReader reader, int version = 0)
@@ -77,7 +73,6 @@ namespace YARG.Core.Game
             FourLaneDrums.Deserialize(reader, version);
             FiveLaneDrums.Deserialize(reader, version);
             ProKeys.Deserialize(reader, version);
-            Common.Deserialize(reader, version);
         }
     }
 }
